Don't use reinterpret cast for generating a unique entitiy id.
id_t::local_id() is used for generating a unique id which gets used in sorting operations.  The local id was being
formed by reinterpret_cast'ing the "this" pointer.  This was not portable as it was causing some bizarre behaviour
in optimized builds.  This commit changes id_t::local_id() to generate a unique id via bit masking.